+++ CHANGELOG.md
@@ -3,3 +3,6 @@
 The format of this file is based on [Keep a Changelog 1.0.0](https://keepachangelog.com/en/1.0.0/).
 
 ## [unreleased]
+### Added
+- `nm_add_gtest`, a function downloading Google Test and
+  adding it to the build